refactor(rql): move canonical SQL AST cluster into reddb-io-rql behind shims
Move the canonical SQL AST — QueryExpr and the whole query-type family
(core.rs), the Phase-2 Expr tree (Expr/Span/UnaryOp/ExprSubquery, ast.rs),
and the query builders (builders.rs) — from reddb-server into reddb-io-rql
(ADR 0053, RQL Phase 2 S4c). Because inherent impls cannot span crates,
QueryExpr::{table,graph,path} (returns builders) and Filter::optimize drag
their dependency-closed helpers with them: filter_optimizer.rs and the whole
sql_lowering.rs lowering module move too. Every file in the cluster depends
only on the AST itself plus reddb-io-types (Value/SqlTypeName/DataType,
MetadataFilter, BinOp, distance/queue_mode/index_hint/catalog leaves), so the
crate graph stays acyclic: reddb-io-rql -> reddb-io-types, with no
reddb-server edge.

The move is byte-faithful: only crate-relative import paths are rewritten
(crate::storage::schema -> reddb_types::types, crate::storage::query::lexer
-> crate::lexer, the intra-cluster module paths -> crate::*). The embedded
AST and filter-optimizer unit tests move verbatim and pass in-crate
(cargo test -p reddb-io-rql --lib: 51 passed).

Server keeps re-export shims at the historical paths so every parser,
planner, evaluator, and runtime call-site resolves unchanged with zero
call-site edits:
  - storage/query/ast.rs            -> pub use reddb_rql::ast::*
  - storage/query/sql_lowering.rs   -> pub use reddb_rql::sql_lowering::*
  - storage/query/filter_optimizer.rs -> pub use reddb_rql::filter_optimizer::*

Two clippy lints the server silenced crate-wide (should_implement_trait on
Filter::not / WithClause::add, dead_code on the unused render_*_label
helpers) are silenced surgically in-crate, matching the repo convention
(cf. 76959147); rql stays clippy/fmt clean.

cargo check --workspace green (reddb-io-server builds against the shims).

Refs #1113
